+++ insonmnia/worker/storage/quotafactory_nonlinux.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package storagequota
+package storage
 
 import (
 	"fmt"
